fix: ipfs base script
